requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.28.1
structlog==24.4.0
